refactor: NonNegativeNumber 패키지 이동 및 접근 제어자 public 으로 변경
