+++ src/test/java/calculator/StringExpression.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import common.numeric.NonNegativeNumber;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
