+++ src/test/java/common/numeric/NonNegativeNumberTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package common.numeric;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
